% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{airports}
\alias{airports}
\title{Airport metadata}
\format{A data frame with columns:
\describe{
 \item{faa}{FAA airport code}
 \item{name}{Usual name of the aiport}
 \item{lat,lon}{Location of airport}
 \item{alt}{Altitude, in feet}
 \item{tz}{Timezone offset from GMT}
 \item{dst}{Daylight savings time zone. A = Standard US DST: starts on the
    second Sunday of March, ends on the first Sunday of November.
    U = unknown. N = no dst.}
}}
\source{
\url{http://openflights.org/data.html}, downloaded 2014-06-27
}
\usage{
airports
}
\description{
This dataset contains useful metadata about airports.
}
\keyword{datasets}

