% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildIndices.R
\name{buildIndices}
\alias{buildIndices}
\title{buildIndices}
\usage{
buildIndices(db)
}
\arguments{
\item{db}{a \code{dplyr} \code{src} or a \code{DBI} connection}
}
\description{
Build indices in the airlines database
}
\examples{
library(dplyr)
if (require(RPostgreSQL)) {
 # must have pre-existing database "airlines"
 db <- src_postgres(host = "localhost", user="postgres", password="postgres", dbname = "airlines")
}

start <- Sys.time()
flights <- tbl(db, "flights")
flights \%>\%
  filter(dest == 'BDL') \%>\%
  collect()

end <- Sys.time()
end - start

# build the indices
\dontrun{
buildIndices(db)

start <- Sys.time()
flights <- tbl(db, "flights")
flights \%>\%
  filter(dest == 'BDL') \%>\%
  collect()

end <- Sys.time()
end - start
}
}

