% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indices.R
\name{dbBuildIndex}
\alias{dbBuildIndex}
\title{Build an index}
\usage{
dbBuildIndex(conn, tablename, cols, indexname = NULL, ...)
}
\arguments{
\item{conn}{A \code{DBIConnection} object, as produced by \code{\link{dbConnect}}}

\item{tablename}{The name of the table you want to build an index on}

\item{cols}{A character vector of column names in the table on which to build
the index}

\item{indexname}{The name of the index to be created (default NULL)}

\item{...}{arguments passed to \code{\link{dbGetQuery}}}
}
\description{
\code{dbBuildIndex} uses the SQL CREATE INDEX syntax to build an index on a
specified column.
}
\examples{
if (require("RSQLite")) {
con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "mtcars", mtcars)
dbGetQuery(con, "SELECT * FROM mtcars")

dbBuildIndex(con, "mtcars", c("cyl", "am"))

dbDisconnect(con)
}
}
\seealso{
Other connection.methods: \code{\link{dbBuildKey}}
}

