% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keys.R
\name{dbBuildKey}
\alias{dbBuildKey}
\title{Build a Key}
\usage{
dbBuildKey(conn, tablename, cols, keyname = NULL, ...)
}
\arguments{
\item{conn}{A \code{DBIConnection} object, as produced by \code{\link{dbConnect}}}

\item{tablename}{The name of the table you want to build an index on}

\item{cols}{A character vector of column names in the table on which to build
the index}

\item{keyname}{The name of the key to be created (default NULL)}

\item{...}{arguments passed to \code{\link{dbGetQuery}}}
}
\description{
\code{dbBuildKey} uses the SQL ALTER TABLE syntax to build a key on a
specified column.
}
\examples{
if (require("RSQLite")) {
con <- dbConnect(RSQLite::SQLite(), ":memory:")

dbWriteTable(con, "mtcars", mtcars)
dbGetQuery(con, "SELECT * FROM mtcars")

dbBuildKey(con, "mtcars", c("cyl", "am"))

dbDisconnect(con)
}
}
\seealso{
Other connection.methods: \code{\link{dbBuildIndex}}
}

