% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_airlines}
\alias{etl_extract.etl_airlines}
\title{etl_extract}
\usage{
\method{etl_extract}{etl_airlines}(obj, year = 2013, month = NULL, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{year}{a year represented as a four-digit integer}

\item{month}{the month represented as an integer}

\item{...}{arguments passed to methods}
}
\description{
a utility function to download a year's worth of airline data
}
\examples{
\dontrun{

library(dplyr)
if (require(RMySQL)) {
# must have pre-existing database "airlines"
db <- src_mysql(user = "mysql", password = "mysql", dbname = "airlines")
}
library(etl)
etl_airlines <- etl_connect("airlines", db, dir = "~/dumps/airlines")
# get one entire year of data
etl_airlines <- etl_extract(etl_airlines, year = 2013)
list.files(etl_airlines$dir)

# get two years worth
lapply(2012:2013, etl_extract, obj = etl_airlines)
list.files(etl_airlines$dir)
}
}
\seealso{
Other etl.functions: \code{\link{etl_cleanup.etl_airlines}}
}

