% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_cleanup.R, R/etl_extract.R, R/etl_init.R, R/etl_load.R, R/etl_transform.R
\name{etl_cleanup.etl_airlines}
\alias{etl_cleanup.etl_airlines}
\alias{etl_extract.etl_airlines}
\alias{etl_init.etl_airlines}
\alias{etl_load.etl_airlines}
\alias{etl_transform.etl_airlines}
\title{ETL functionality for the airlines database}
\usage{
\method{etl_cleanup}{etl_airlines}(obj, delete_zip = FALSE,
  delete_csv = FALSE, ...)

\method{etl_extract}{etl_airlines}(obj, year = 2015, months = 1:12, ...)

\method{etl_init}{etl_airlines}(obj, ...)

\method{etl_load}{etl_airlines}(obj, year = 2015, months = 1:12, ...)

\method{etl_transform}{etl_airlines}(obj, year = 2015, months = 1:12, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{delete_zip}{should the ZIP files we downloaded be deleted?}

\item{delete_csv}{should the CSV files we downloaded by deleted?}

\item{...}{arguments passed to methods}

\item{year}{a year represented as a four-digit integer}

\item{months}{a vector of integers representing the months}
}
\description{
Perform ETL operations on the airlines database.
}
\details{
If a \code{year} and/or \code{month} is specified, then
only flight data from matching months is used.

This will initialize and populate the database.
}
\examples{
require(dplyr)
require(etl)
\dontrun{
if (require(RPostgreSQL)) {
# must have pre-existing database "airlines"
db <- src_postgres(host = "localhost", user = "postgres", dbname = "airlines")
}
airlines <- etl_connect("airlines", db)
etl_init(airlines)

if (require(RMySQL)) {
# must have pre-existing database "airlines"
db <- src_mysql(user = "r-user", password = "mypass", dbname = "airlines")
}
airlines <- etl("airlines", db)
etl_init(airlines)

etl_update(airlines, year = 2013, month = 6)
}
\dontrun{
library(dplyr)
if (require(RMySQL)) {
# must have pre-existing database "airlines"
db <- src_mysql(user = "r-user", password = "mypass", dbname = "airlines")
}
library(etl)
etl_airlines <- etl_connect("airlines", db, dir = "~/dumps/airlines")
# get one entire year of data
etl_airlines <- etl_airlines \%>\%
  etl_extract(year = 2013) \%>\%
  etl_transform(month = 6) \%>\%
  etl_load(year = 2013, month = 6)
}

#' # check the results
\dontrun{
flights.db <- tbl(db, "flights")
flights.db \%>\%
  group_by(year, origin) \%>\%
  summarise(N = n(), min.month = min(month), max.month = max(month)) \%>\%
  arrange(desc(N))
}
}
\seealso{
\code{\link[etl]{etl_cleanup}}

\code{\link[etl]{etl_init}}
}

