% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_load.R
\name{etl_load.etl_airlines}
\alias{etl_load.etl_airlines}
\title{etl_load}
\usage{
\method{etl_load}{etl_airlines}(obj, year = NULL, month = NULL, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{year}{four-digit integer indicating the year to upload}

\item{month}{integer indicating the month to upload}

\item{...}{arguments passed to methods}
}
\description{
a utility function to push a data frame to a DB.
}
\examples{
\dontrun{
library(dplyr)
if (require(RMySQL)) {
# must have pre-existing database "airlines"
db <- src_mysql(user = "mysql", password = "mysql", dbname = "airlines")
}
library(etl)
etl_airlines <- etl_connect("airlines", db, dir = "~/dumps/airlines")
# get one entire year of data
etl_airlines <- etl_airlines \%>\%
  etl_extract(year = 2013) \%>\%
  etl_transform(month = 6) \%>\%
  etl_load(year = 2013, month = 6)
}

#' # check the results
\dontrun{
flights.db <- tbl(db, "flights")
flights.db \%>\%
  group_by(year, origin) \%>\%
  summarise(N = n(), min.month = min(month), max.month = max(month)) \%>\%
  arrange(desc(N))
}
}

