% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_process.R
\name{etl_process.etl_airlines}
\alias{etl_process.etl_airlines}
\title{etl_process}
\usage{
\method{etl_process}{etl_airlines}(obj, year = NULL, month = NULL, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{year}{a year represented as a four-digit integer}

\item{month}{the month represented as an integer}

\item{...}{arguments passed to methods}
}
\description{
a utility function to process a single month's worth of airline data
}
\examples{
\dontrun{

library(dplyr)
if (require(RMySQL)) {
# must have pre-existing database "airlines"
db <- src_mysql(user = "mysql", password = "mysql", dbname = "airlines")
}
library(etl)
etl_airlines <- etl_connect("airlines", db, dir = "~/dumps/airlines")
# get one entire year of data
etl_airlines <- etl_scrape(etl_airlines, year = 2013)
list.files(etl_airlines$dir)
etl_process(etl_airlines, month = 6)
list.files(etl_airlines$dir)
}
}

