% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flights}
\alias{flights}
\title{Flights}
\format{Data frame with columns
\describe{
\item{year,month,day}{Date of departure}
\item{dep_time,arr_time}{Actual departure and arrival times, local tz.}
\item{sched_dep_time,sched_arr_time}{Scheduled departure and arrival times, local tz.}
\item{dep_delay,arr_delay}{Departure and arrival delays, in minutes.
  Negative times represent early departures/arrivals.}
\item{hour,minute}{Time of scheduled departure broken into hour and minutes.}
\item{carrier}{Two letter carrier abbreviation. See \code{\link{carriers}}
  to get name}
\item{tailnum}{Plane tail number}
\item{flight}{Flight number}
\item{origin,dest}{Origin and destination. See \code{\link{airports}} for
  additional metadata.}
\item{air_time}{Amount of time spent in the air}
\item{distance}{Distance flown}
\item{time_hour}{Scheduled date and hour of the flight as a \code{POSIXct} date.
  Along with \code{origin}, can be used to join flights data to weather data.}
}}
\source{
RITA, Bureau of transportation statistics,
 \url{http://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}
}
\usage{
flights
}
\description{
On-time data for all flights.
}
\keyword{datasets}

