% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMonth.R
\name{getMonth}
\alias{getMonth}
\title{getMonth}
\usage{
getMonth(year = 2013, month = 1, temp.dir = tempdir())
}
\arguments{
\item{year}{a year represented as a four-digit integer}

\item{month}{the month represented as an integer}

\item{temp.dir}{a directory to store the raw montly data files to be downloaded.}
}
\description{
a utility function to download a single month's worth of airline data
}
\examples{
library(dplyr)
\dontrun{
May <- getMonth(year = 2013, month = 5)
head(May)
dim(May)
May \%>\%
  group_by(year, origin) \%>\%
  summarise(N = n(), min.month = min(month), max.month = max(month)) \%>\%
  arrange(desc(N))
}
}

