% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycflights13-package.R
\docType{package}
\name{nycflights13-package}
\alias{nycflights13}
\alias{nycflights13-package}
\title{nycflights13: Flights that Departed NYC in 2013}
\description{
Airline on-time data for all flights departing NYC in 2013. Also includes useful 'metadata' on airlines, airports, weather, and planes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nycflights13.tidyverse.org}
  \item \url{https://github.com/tidyverse/nycflights13}
  \item Report bugs at \url{https://github.com/tidyverse/nycflights13/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.com}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
