% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\alias{pushFlights}
\alias{pushFlightsYear}
\title{push}
\usage{
push(db, df, tablename, ...)

pushFlights(db, df, tablename = "flights", append = TRUE,
  row.names = FALSE)

pushFlightsYear(db, year, ...)
}
\arguments{
\item{db}{a \code{dplyr} \code{src} or a \code{DBI} connection}

\item{df}{a \code{tbl_df} to push to the database server}

\item{tablename}{the name of the table in the database to which you want to
push the data. The default is "flights".}

\item{...}{arguments passed to \code{getMonth} or \code{dbWriteTable}.}

\item{year}{a year represented as a four-digit integer}
}
\description{
a utility function to push a data frame to a DB.
}
\examples{
#' library(RMySQL)
# must have pre-existing database "airlines"
db <- src_mysql(host = "localhost", user="bbaumer", password="fakepass", dbname = "airlines")

library(RPostgreSQL)
# must have pre-existing database "airlines"
db <- src_postgres(host = "localhost", user="postgres", password="postgres", dbname = "airlines")

#' # save files in a specific location
\dontrun{
feb <- getMonth(year = 2013, month = 2, temp.dir = "~/dumps")
pushFlights(db, feb)
pushYear(db, year = 2013, temp.dir = "~/dumps")
}

#' # check the results
\dontrun{
flights.db <- tbl(db, "flights")
flights.db \%>\%
  group_by(year, origin) \%>\%
  summarise(N = n(), min.month = min(month), max.month = max(month)) \%>\%
  arrange(desc(N))
}
}

