% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{trim}
\usage{
trim()
}
\description{
Enforce minimal referential integrity on the flights DB.
}
\examples{
library(RPostgreSQL)
# must have pre-existing database "airlines"
db <- src_postgres(host = "localhost", user="postgres", password="postgres", dbname = "airlines")

#' # save files in a specific location
\dontrun{
feb <- getMonth(year = 2013, month = 2, temp.dir = "~/dumps")
pushFlights(db, feb)
pushYear(db, year = 2013, temp.dir = "~/dumps")
}

#' # check the results
\dontrun{
flights.db <- tbl(db, "flights")
flights.db \%>\%
  group_by(year, origin) \%>\%
  summarise(N = n(), min.month = min(month), max.month = max(month)) \%>\%
  arrange(desc(N))
}
}

