% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{airlines}
\alias{airlines}
\title{Airline names.}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
http://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236
}
\usage{
airlines
}
\description{
Look up airline names from their carrier codes.
}
\examples{
airlines
}
\keyword{datasets}

