% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/airport.R
\docType{data}
\name{airports}
\alias{airports}
\title{Airport metadata}
\format{A data frame with columns:
\describe{
 \item{faa}{FAA airport code}
 \item{name}{Usual name of the aiport}
 \item{lat,lon}{Location of airport}
 \item{alt}{Altitude, in feet}
 \item{tz}{Timezone offset from GMT}
 \item{dst}{Daylight savings time zone. A = Standard US DST: starts on the
    second Sunday of March, ends on the first Sunday of November.
    U = unknown. N = no dst.}
}}
\source{
\url{http://openflights.org/data.html}, downloaded 2014-06-27
}
\usage{
airports
}
\description{
Useful metadata about airports.
}
\examples{
if (require("dplyr")) {
airports

airports \%>\% mutate(dest = faa) \%>\% semi_join(flights)
flights \%>\% anti_join(airports \%>\% mutate(dest = faa))
airports \%>\% mutate(origin = faa) \%>\% semi_join(flights)

}
}
\keyword{datasets}

