% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carriers}
\alias{carriers}
\title{Carriers}
\format{Data frame with columns
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}}
\source{
\url{http://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}
}
\usage{
carriers
}
\description{
Look up airline names from their carrier codes.
}
\keyword{datasets}

