% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_cleanup.R
\name{etl_cleanup.etl_airlines}
\alias{etl_cleanup.etl_airlines}
\title{Cleanup after data loaded into DB}
\usage{
\method{etl_cleanup}{etl_airlines}(obj, delete_zip = FALSE,
  delete_csv = FALSE, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
Cleanup after data loaded into DB
}
\examples{
\dontrun{
require(RPostgreSQL)
# connect directly
require(dplyr)
db <- src_postgres("mtcars", user = "postgres", host = "localhost")
etl_cars <- etl_connect("mtcars", db)
etl_cars \%>\%
 etl_create() \%>\%
 etl_cleanup()
}
}
\seealso{
Other etl.functions: \code{\link{etl_scrape.etl_airlines}}
}

