% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/etl_init.R
\name{etl_init.etl_airlines}
\alias{etl_init.etl_airlines}
\title{etl_init}
\usage{
\method{etl_init}{etl_airlines}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{...}{arguments passed to methods}
}
\description{
Initialize the database.
}
\details{
This will initialize the database.
}
\examples{
library(dplyr)
library(etl)

\dontrun{
if (require(RPostgreSQL)) {
# must have pre-existing database "airlines"
db <- src_postgres(host = "localhost", user = "postgres", dbname = "airlines")
}
etl_airlines <- etl_connect("airlines", db)
etl_init(etl_airlines)

if (require(RMySQL)) {
# must have pre-existing database "airlines"
db <- src_mysql(user = "mysql", password = "mysql", dbname = "airlines")
}
etl_airlines <- etl_connect("airlines", db)
etl_init(etl_airlines)
}
}
\seealso{
\code{\link[etl]{etl_init}}
}

